% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{makeTestSet}
\alias{makeTestSet}
\title{makeTestSet}
\usage{
makeTestSet(data, variables, avgfun)
}
\arguments{
\item{data}{A dataset}

\item{variables}{A list of variables. The names should correspond to the names of variables in the dataset, while the values should be the values you want the variables to have.}

\item{avgfun}{Function to produce placeholder values for variables not included in the variables list.}
}
\description{
Create a test set from a dataset. The resulting dataset
contains every permutational combination of variable values from the
variables list. This dataset is useful for testing counterfactuals.
}
