% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim}
\alias{sim}
\title{sim}
\usage{
sim(data, model, n = 1000, vcovFn = vcov, interval = c(0.025, 0.975))
}
\arguments{
\item{data}{A dataset}

\item{model}{Model to simulate with. Must support vcov and coef}

\item{n}{Number of draws}

\item{vcovFn}{function that produces a VCOV}
}
\description{
Simulates outcomes from a model with parameter uncertainties.
}
